requirements.txt
streamlit
numpy
pandas
Pillow
opencv-python
plotly
matplotlib
insightface
onnxruntime